Add test case of describe_instance_types

+++ src/compute_api/include/model/instance_type.hpp
@@ -31,13 +31,13 @@ namespace model
 	class instance_type
 	{
 	private:
-		float vcpus;
+		unsigned vcpus;
 		float ram;
 		string id;
 
 	public:
 		///Constructor
-		instance_type(float vcpus_, float ram_, const string& id_):
+		instance_type(unsigned vcpus_, float ram_, const string& id_):
       vcpus(vcpus_),
 			ram(ram_),
 			id(id_)
@@ -48,7 +48,7 @@ namespace model
 		/**
 			\return const float
 		*/
-		float get_vcpus() const
+		unsigned get_vcpus() const
 		{
 			return vcpus;
 		}
